// src/types/user.ts
//Типы для userReducer

//Перечисления с типами action
export enum UserActionTypes {
    FETCH_USERS = 'FETCH_USERS',
    FETCH_USERS_SUCCESS = 'FETCH_USERS_SUCCESS',
    FETCH_USERS_ERROR = 'FETCH_USERS_ERROR',
}

//Типизации для каждого action 
interface FetchUserAction {
    type: UserActionTypes.FETCH_USERS;
}

interface FetchUserSuccessAction {
    type: UserActionTypes.FETCH_USERS_SUCCESS;
    payload: any;
}

interface FetchUserErrorAction {
    type: UserActionTypes.FETCH_USERS_ERROR;
    payload: string;
}

//Типизация состояния error
export interface UserState {
    users: any[];
    loading: boolean;
    error: null | string;
}

//Объединение action воедино, универсальная типизация action
export type UserAction = FetchUserAction | FetchUserSuccessAction | FetchUserErrorAction;